fixed exception when cancelling delete video prompt

+++ capstone_video_analyzer/lib/pages/video_player_page.dart
@@ -34,7 +34,7 @@ class VideoPlayerPage extends StatelessWidget {
               ],
             );
           });
-      if (deleteSelected) {
+      if (deleteSelected != null && deleteSelected) {
         onDeleteVideo(url);
         Navigator.pop(context);
       }
